Replace #cmakedefine by #define for G4VERSION_(MAJOR/MINOR/PATCH) variables.
G4VERSION_(MAJOR/MINOR/PATCH) variables are always defined and not of boolean type so #cmakedefine is not suitable in these cases.
For example, G4VERSION_PATCH can be equal to zero thus leading to it's replacement by an #undef.
As main function uses these C++ macros values, these must be always defined.