correct root flags in AnalyzeCT.cpp compilation
The order in which the flags appear after g++ matters. The "compile" flags should go before the source file, the linking flags should go after, so you cannot combine them. Maybe the combined blurb was working with older versions of g++ and/or ROOT, but with g++ 9 and ROOT 6 it seems necessary to keep the compile flags and linking flags separate.